Generate merged code coverage reports with pycobertura
Cobertura is a coverage tool for Java that outputs reports in XML.
The tools that we use for Python and C coverage respectively,
coverage.py and gcovr, can both output the same format. So we can
use tools for Cobertura to generate merged, multilingual coverage
reports.

This would be a lot prettier if we could use coveralls or
codecov.io, which support multilingual test coverage. However,
those products don't integrate with git.ligo.org (or at least, they
don't integrate for free).