Interpolate amplitude and phase of SNR time series, not real and imaginary parts
Change BAYESTAR's strategy for evaluating SNR time series from
Catmull-Rom interpolation of the real and imaginary parts to
Catmull-Rom interpolation of the amplitude and phase. The old
interpolation method could produce oscillatory artifacts in the SNR
amplitude if the data are nearly critically sampled, as is the case for
early-warning BNS events. The new interpolation method is immune to
this kind of artifact, and also has much faster convergence as a
function of sample rate.