Add options to ligo-skymap-plot to control the --radec marker.
The following options are added that specify shape, color, edge color
and size of the marker used with the --radec option that already
existed:
--marker to specify the symbol
--marker-color for the color
--marker-ecolor for the edge color (outside line of the marker)
--marker-size for the size

Defaults are implemented so that the behaviour does not change when NOT
specifying any of these new options. I.e. by default, --radec plots a
size 10 star ('*') marker in white with a black outline.