Handle edge cases for 0, +inf in {conditional,marginal}_cdf
Ensure that `ligo.skymap.distance.conditional_cdf` and
`ligo.skymap.distance.marginal_cdf` correctly handle corner cases
for extreme values: both now return 0 for distances that are less
than or equal to 0, and 1 for a distance of positive infinity
(assuming that the ``distnorm`` argument does indeed normalize the
distribution).

Fixes #51.